+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes qsf-integration project will be documented in this file.
 ## [Unreleased]
 ### Added
 - make it possible to set a custom searchQuery object in the elastic query transformer
+- implement tracking for range values
 
 
 ### Changed